requirements.txt
anyio==4.6.2
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
pillow==10.4.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
